refactor(desktop): extract useKeymap (chord map + VIM ops) from App.tsx (F5.12)
The keyboard action layer moves into useKeymap: the chord->action lookup
built from the config keymap, plus the VIM range-operation state machine
(a3a/d2d/t5t/l2l) with runVimSingle/runVimRange/clearVimRange and the
vimRange ref. Consumed by the keydown handler via ctx; verbatim move.

App.tsx 1,351 -> 1,203; useKeymap 187. unit 61 / e2e 39 green, tsc +
build clean.
